pyproject.toml
[project]
name = "syntrack"
version = "0.1.0.dev0"
description = "Genome synteny visualization tool"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Petr Novak" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.5",
    "numpy>=2.0",
    "polars>=1.0",
    "typer>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-benchmark>=4.0",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.11",
    "types-PyYAML",
]

[project.scripts]
syntrack = "syntrack.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["syntrack"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: requires example_data/ symlinks; slow",
]
